+++ internal/align/terraform_test.go
@@ -13,6 +13,7 @@ import (
 func TestTerraformAttributeOrderAndBlocks(t *testing.T) {
 	src := []byte(`terraform {
   backend "s3" {}
+  experiments = ["a"]
   required_providers {}
   required_version = ">= 1.2.0"
   other = 1
@@ -24,6 +25,7 @@ func TestTerraformAttributeOrderAndBlocks(t *testing.T) {
 	got := string(file.Bytes())
 	exp := `terraform {
   required_version = ">= 1.2.0"
+  experiments      = ["a"]
 
   required_providers {}