+++ tests/cases/terraform/experiments/out.tf
@@ -1,0 +1,11 @@
+terraform {
+  required_version = ">= 1.2.0"
+  experiments      = ["a"]
+
+  required_providers {}
+
+  backend "s3" {}
+
+  cloud {}
+  other = 1
+}